requirements.txt
fastapi
uvicorn
boto3
pandas
numpy
faiss-cpu
